requirements.txt
# SNP Printer Service Dependencies
flask==3.0.0
python-escpos==3.1
Pillow==10.1.0
Werkzeug==3.0.1
